Implements Basic Interactor
This interactor doesn't have an output delegate defined.

The rational behind this is to promote an API where the caller
passes a completion handler or the callee returns some kind of
Promise/Future instead of using delegates.

Using delegates adds unnecessary state management to the presenter
since it's hard to tell which output corresponds to which input.